Time: 609 ms (26.08%), Space: 6 MB (47.83%) - LeetHub

+++ 1680-concatenation-of-consecutive-binary-numbers/1680-concatenation-of-consecutive-binary-numbers.cpp
@@ -12,7 +12,7 @@ public:
     
     int concatenatedBinary(int n) {
         unsigned long long int ans = 0;
-        unsigned long long int M = 1e9+7;
+         int M = 1e9+7;
         
         for(int i=1;i<=n;i++){
             int bits[17];
